requirements.txt
fastapi
pydantic
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
alembic
psycopg2
python-dotenv
pytest
requests
